+++ citizens/config.py
@@ -14,7 +14,7 @@ class Settings(BaseSettings):
 
     # Provided by AppAPI / the container environment
     app_id: str = "citizens"
-    app_version: str = "0.6.0-beta.9"
+    app_version: str = "0.6.0-beta.10"
     app_host: str = "0.0.0.0"
     app_port: int = 23000
     app_secret: str = ""